+++ src/m_swap.h
@@ -44,8 +44,8 @@
 #define LONG(x)	((INT32)(x))
 #endif
 #else // no need for swapping
-#define SHORT(x) (x)
-#define LONG(x)	(x)
+#define SHORT(x) ((INT16)x)
+#define LONG(x)	((INT32)x)
 #endif // ifndef _NDS
 
 #endif
